requirements.txt
acme==1.19.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argcomplete==1.12.0
attrs==19.3.0
Babel==2.8.1
Beaker==1.10.0
beautifulsoup4==4.9.3
blivet==3.3.0
blivet-gui==2.2.1
Brlapi==0.8.0
cached-property==1.5.1
certbot==1.19.0
certifi==2024.7.4
cffi==1.14.1
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==8.1.7
ConfigArgParse==1.2.3
configobj==5.0.6
cryptography==3.2.1
cupshelpers==1.0
dasbus==1.4
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.7.0rc2
distlib==0.3.1
distro==1.5.0
dnspython==2.6.1
docopt==0.6.2
docutils==0.16
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
feedparser==5.2.1
filelock==3.0.12
freezegun==0.3.12
fros==1.1
gevent==20.6.1
gpg==1.14.0
greenlet==0.4.16
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.5
httplib2==0.18.1
httptools==0.6.1
httpx==0.27.0
humanize==0.5.1
idna==3.7
isodate==0.6.0
Jinja2==2.11.3
josepy==1.8.0
jwt==1.3.1
langtable==0.0.56
libcomps==0.1.18
lxml==4.5.1
Mako==1.1.3.dev0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mock==3.0.5
mysql-connector-python==8.4.0
nftables==0.1
num2words==0.5.10
numpy==1.19.4
ofxparse==0.18
olefile==0.46
openpyxl==3.1.5
openshot-qt==2.5.1
orjson==3.10.5
packaging==20.4
parsedatetime==2.5
passlib==1.7.4
Paste==3.4.0
pexpect==4.8.0
pid==2.2.3
pillow==10.3.0
ply==3.11
polib==1.1.0
productmd==1.33
psutil==5.7.2
psycopg2==2.8.6
ptyprocess==0.6.0
pwquality==1.4.4
pyasn1==0.4.8
pycairo==1.18.2
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.43.0.5
pydantic==2.8.2
pydantic_core==2.20.1
pydot==1.4.1
pyenchant==3.1.1
Pygments==2.18.0
PyGObject==3.38.0
PyJWT==2.8.0
pykickstart==3.29
pymssql==2.3.0
PyMySQL==1.1.1
pyodbc==5.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyparted==3.11.5
PyPDF2==1.26.0
PyQt5==5.15.0
PyQt5_sip==4.19.24
pyRFC3339==1.1
pyserial==3.4
PySocks==1.7.1
python-augeas==0.5.0
python-dateutil==2.8.1
python-dotenv==1.0.1
python-meh==0.50
python-multipart==0.0.9
python-stdnum==1.14
pytz==2020.5
pyudev==0.22.0
pyusb==1.0.2
pyxdg==0.26
PyYAML==6.0.1
pyzmq==19.0.0
qrcode==6.1
reportlab==3.5.57
requests==2.24.0
requests-file==1.4.3
requests-ftp==0.3.1
requests-toolbelt==0.9.1
rich==13.7.1
rpm==4.16.1.3
selinux==3.1
sepolicy==3.1
setools==4.3.0
shellingham==1.5.4
simpleaudio==1.0.4
simpleline==1.7
six==1.15.0
slip==0.6.4
slip.dbus==0.6.4
sniffio==1.3.1
sos==4.0
soupsieve==2.0.1
SQLAlchemy==2.0.31
sqlalchemy-pyodbc-mssql==0.1.1
starlette==0.37.2
systemd-python==234
Tempita==0.5.1
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==1.25.8
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.0.30
vobject==0.9.6.1
watchfiles==0.22.0
websockets==12.0
Werkzeug==1.0.1
xlrd==1.2.0
xlwt==1.3.0
zeep==3.4.0
zope.component==4.3.0
zope.event==4.2.0
zope.interface==5.1.0
